rename subpackage back to `common`
(was temporarily changed to `input`)
